Fix competitor count subqueries: cast bigint to int
PostgreSQL COUNT(*) returns bigint which Drizzle serializes as a string.
Cast to int so the JSON response contains actual numbers.

+++ src/app/api/competitors/route.ts
@@ -13,10 +13,10 @@ export async function GET() {
       linkedinSlug: competitors.linkedinSlug,
       notes: competitors.notes,
       createdAt: competitors.createdAt,
-      clientCount: sql<number>`(SELECT COUNT(*) FROM competitor_clients WHERE competitor_id = ${competitors.id})`,
-      employeeCount: sql<number>`(SELECT COUNT(*) FROM competitor_employees WHERE competitor_id = ${competitors.id})`,
-      postCount: sql<number>`(SELECT COUNT(*) FROM competitor_posts WHERE competitor_id = ${competitors.id})`,
-      engagementCount: sql<number>`(SELECT COUNT(DISTINCT engager_profile_url) FROM competitor_post_engagement WHERE competitor_id = ${competitors.id})`,
+      clientCount: sql<number>`(SELECT COUNT(*)::int FROM competitor_clients WHERE competitor_id = ${competitors.id})`,
+      employeeCount: sql<number>`(SELECT COUNT(*)::int FROM competitor_employees WHERE competitor_id = ${competitors.id})`,
+      postCount: sql<number>`(SELECT COUNT(*)::int FROM competitor_posts WHERE competitor_id = ${competitors.id})`,
+      engagementCount: sql<number>`(SELECT COUNT(DISTINCT engager_profile_url)::int FROM competitor_post_engagement WHERE competitor_id = ${competitors.id})`,
     })
     .from(competitors)
     .orderBy(competitors.name);
